Audit > Update request filtering and file handling
• Update audit driver to ignore requests matching block list patterns
• Enhance file event resource generation with thumb and transcode options
• Refactor transcode job: remove redundant options parameter via property promotion
• Cast timeout value to integer in PDF-to-images transcoder

+++ src/Services/TranscodeFile/PdfToImagesTranscoder.php
@@ -15,7 +15,7 @@ class PdfToImagesTranscoder extends FileTranscoderAbstract implements FileTransc
 
 	public function getTimeout(StoredFile $storedFile): int
 	{
-		return $this->timeEstimate($storedFile) / 1000 * 3;
+		return (int)($this->timeEstimate($storedFile) / 1000 * 3);
 	}
 
 	public function timeEstimate(StoredFile $storedFile): int
